fixed circular import error with electrolyzer cost tools
